+++ src/model/OrangeJuice.java
@@ -39,7 +39,7 @@ public class OrangeJuice implements BeverageComponent {
                      sugars = result.getInt("sugars");
                     carbohydrates = result.getInt("carbohydrates");
                     calories= result.getInt("calories");
-                    price = result.getInt("price");
+                    price = result.getDouble("price");
                     count = result.getInt("count");
                 }
